Support downloading external sources
Since Godot doesn't support streaming video in real time it needs to be downloaded completely before playback.
In that case the VideoTexture will return a texture showing the text "Buffering.." until the video is downloaded.
Once the download is finished if playing is set to true it'll start playing the video